Update
Added the 'eat', 'eaten', and 'table' commands.
The former two are similar to points.
The latter is more interesting, and allows dynamic access to tables of
entries that can be added and queried.
This can be used to have for example a quote database that users can
query for a random quite, as seen in other bots.

Added a database backup.
The power was cut off from my machine, and apparently it was just saving
the DB, and it was all lost.
Having an additional (one step older) backup should help. Duh.