Update
Timer now sends itself as the first argument to its callback.

Merged the add/edit/remove commands to the commands/intervals, which
follow the syntax of the table command.

Added an ignore list per-channel, and the ignore command to dynamically
use it from chat.
The bot will not listen to commands from ignored users.
The ignores are timed, and update automatically (and are saved in the DB
of course).