Update
Added TwitchApi.getRecentMessages().

Channels now store their message queues which are polled by the bot.
This allows for the mute command to actually mute current messages
rather than only future ones.

Added channel.settings.commandsDelay, which is a number that determines
the speed at which commands are accepted.
This makes it possible to avoid command spam.
The default for now is 500ms.