Update
Hopefully fixed the issue where the bot wouldn't rejoin channels after a
reconnection, and generally improved the behavior of reconneting.

Added the RECONNECT command from Twitch, which will, quite obviously,
reconnect the bot.

The bot 'connected' event is now sent when the bot connects to the IRC
chat, rather than when the socket is connected.

Added the bot 'joined' event that is send together with a channel name,
whenever a channel is joined.